feat: добавлены шаблоны для процессных разрезов
Был выполнен рефакторинг структуры файлов в директории indicatorsFormulas
