+++ src/indicatorsFormulas/dimension/timeTemplates.ts
@@ -1,0 +1,51 @@
+import { convertFiltersToFormula } from "../../calculators/utils/filters";
+import { EWidgetIndicatorValueModes, type TWidgetIndicatorTimeValue } from "../../indicators";
+import { fillTemplateString } from "../shared";
+import { dimensionTemplateFormulas, type EDimensionTemplateNames } from "./baseTemplates";
+
+/** Шаблоны процессных метрик разреза с режимами START_TIME/END_TIME */
+export const timeTemplates = (() => {
+  const generateTemplates = (innerTemplate: string) => {
+    const templates = {} as Record<EDimensionTemplateNames, string>;
+
+    for (const key in dimensionTemplateFormulas) {
+      templates[key as EDimensionTemplateNames] = fillTemplateString(
+        dimensionTemplateFormulas[key as EDimensionTemplateNames],
+        { columnFormula: innerTemplate }
+      );
+    }
+
+    return templates;
+  };
+
+  return {
+    [EWidgetIndicatorValueModes.START_TIME]: generateTemplates(
+      "process(minIf({eventTimeFormula}, {eventNameFormula} = '{eventName}'{filters}), {caseCaseIdFormula})"
+    ),
+    [EWidgetIndicatorValueModes.END_TIME]: generateTemplates(
+      "process(maxIf({eventTimeFormula}, {eventNameFormula} = '{eventName}'{filters}), {caseCaseIdFormula})"
+    ),
+  };
+})();
+
+/** На основе значения режимов START_TIME/END_TIME подготовить параметры для подстановки в шаблонную формулу */
+export const prepareTimeParams = (value: TWidgetIndicatorTimeValue) => {
+  if (
+    !value.eventName ||
+    !value.caseCaseIdFormula ||
+    !value.eventNameFormula ||
+    !value.processName ||
+    !value.templateName ||
+    !value.eventTimeFormula
+  ) {
+    return;
+  }
+
+  return {
+    eventTimeFormula: value.eventTimeFormula,
+    eventNameFormula: value.eventNameFormula,
+    caseCaseIdFormula: value.caseCaseIdFormula,
+    filters: convertFiltersToFormula(value.filters),
+    eventName: value.eventName,
+  };
+};
