+++ src/indicatorsFormulas/eventMeasure/templates.ts
@@ -1,0 +1,9 @@
+export enum EEventMeasureTemplateNames {
+  eventsCount = "eventsCount",
+  reworksCount = "reworksCount",
+}
+
+export const eventMeasureTemplateFormulas = {
+  [EEventMeasureTemplateNames.eventsCount]: `count()`,
+  [EEventMeasureTemplateNames.reworksCount]: `count() - uniqExact({caseCaseIdFormula})`,
+} as const;
