+++ src/indicatorsFormulas/shared/fillTemplateString.ts
@@ -1,0 +1,3 @@
+export function fillTemplateString(templateString: string, params: Record<string, any>) {
+  return templateString.replace(/\{(.*?)\}/g, (_, key) => params[key] ?? "");
+}
